+++ orchestrator/src/client/pages/DesignResumePage.tsx
@@ -1,64 +1,34 @@
-import * as api from "@client/api";
+import {
+  DESIGN_RESUME_NAV_GROUPS,
+  DESIGN_RESUME_PAGE_MAIN_CLASS_NAME,
+  getDesignResumeSectionIcon,
+  getDeviceLayout,
+} from "@client/components/design-resume/constants";
+import { DesignResumeDock } from "@client/components/design-resume/DesignResumeDock";
 import { DesignResumePreviewPanel } from "@client/components/design-resume/DesignResumePreviewPanel";
 import { DesignResumeRail } from "@client/components/design-resume/DesignResumeRail";
 import { ItemDialog } from "@client/components/design-resume/ItemDialog";
+import {
+  asArray,
+  asRecord,
+  getByPath,
+  getDesignResumeDialogItem,
+  toText,
+} from "@client/components/design-resume/utils";
 import { PageHeader, PageMain } from "@client/components/layout";
+import { SectionWorkspacePanel } from "@client/components/section-workspace/SectionWorkspace";
+import { useDesignResumeStudio } from "@client/hooks/useDesignResumeStudio";
+import type { DesignResumeJson } from "@shared/types";
 import {
-  type SectionWorkspaceBadge,
-  type SectionWorkspaceGroup,
-  SectionWorkspacePanel,
-} from "@client/components/section-workspace/SectionWorkspace";
-import { useDesignResume } from "@client/hooks/useDesignResume";
-import { useSettings } from "@client/hooks/useSettings";
-import { useTracerReadiness } from "@client/hooks/useTracerReadiness";
-import type {
-  DesignResumeDocument,
-  DesignResumeJson,
-  PdfRenderer,
-  TypstTheme,
-} from "@shared/types";
-import { PDF_RENDERER_LABELS, TYPST_THEME_LABELS } from "@shared/types";
-import { useQueryClient } from "@tanstack/react-query";
-import {
-  type MotionValue,
-  motion,
-  useMotionValue,
-  useSpring,
-  useTransform,
-} from "framer-motion";
-import {
-  Award,
-  BookOpen,
-  BriefcaseBusiness,
   Download,
-  Eye,
   FileDown,
-  FileText,
-  Folder,
-  GraduationCap,
-  GripVertical,
-  HeartHandshake,
-  ImageIcon,
   Import,
-  Languages,
-  Link2,
   ListPlus,
-  type LucideIcon,
   MoreHorizontal,
   PenSquare,
-  Quote,
-  ScrollText,
-  Sparkles,
-  Trophy,
-  UserRound,
-  Wrench,
 } from "lucide-react";
 import type React from "react";
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { Navigate, useNavigate, useParams } from "react-router-dom";
-import { toast } from "sonner";
-import { showErrorToast } from "@/client/lib/error-toast";
-import { downloadDesignResumePdf } from "@/client/lib/private-pdf";
+import { Navigate, useNavigate } from "react-router-dom";
 import {
   AlertDialog,
   AlertDialogAction,
@@ -83,1210 +53,57 @@ import {
   SheetHeader,
   SheetTitle,
 } from "@/components/ui/sheet";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
 import { bucketCount, trackProductEvent } from "@/lib/analytics";
 import { cn } from "@/lib/utils";
-import {
-  ITEM_DEFINITIONS,
-  type ItemDefinition,
-} from "../components/design-resume/definitions";
-import {
-  asArray,
-  asRecord,
-  fileToDataUrl,
-  getByPath,
-  getDesignResumeDialogItem,
-  getSectionOrder,
-  makeDownload,
-  REORDERABLE_SECTION_KEYS,
-  toText,
-} from "../components/design-resume/utils";
 import { formatUserFacingError } from "../lib/error-format";
-import { queryKeys } from "../lib/queryKeys";
-
-type DesignResumeSectionId = string;
-type DesignResumeGroupId = "profile" | "sections";
-type DesignResumeNavItem = {
-  id: DesignResumeSectionId;
-  label: string;
-  description: string;
-  icon: LucideIcon;
-  sectionId?: DesignResumeSectionId | null;
-};
-type DesignResumeIconGroupId = "preview" | DesignResumeGroupId;
-type DesignResumeNavGroup = {
-  id: DesignResumeIconGroupId;
-  label: string;
-  items: DesignResumeNavItem[];
-};
-type DesignResumeMobileView = "edit" | "preview";
-
-const DESIGN_RESUME_PAGE_MAIN_CLASS_NAME =
-  "flex min-h-0 flex-1 flex-col space-y-0 overflow-hidden py-3 pb-3";
-
-const SECTION_ICON_BY_ID: Record<string, LucideIcon> = {
-  profiles: Link2,
-  experience: BriefcaseBusiness,
-  education: GraduationCap,
-  projects: Folder,
-  skills: Wrench,
-  languages: Languages,
-  interests: Sparkles,
-  awards: Trophy,
-  certifications: Award,
-  publications: ScrollText,
-  volunteer: HeartHandshake,
-  references: Quote,
-};
-
-const DESIGN_RESUME_PROFILE_SECTIONS: SectionWorkspaceGroup<
-  DesignResumeGroupId,
-  DesignResumeSectionId
->["items"] = [
-  {
-    id: "basics",
-    label: "Contact",
-    description: "Name, headline, and contact details.",
-    searchTerms: ["basics", "headline", "email", "phone", "location"],
-  },
-  {
-    id: "summary",
-    label: "Summary",
-    description:
-      "Short intro shown near the top of your resume. Rewritten during Job Tailoring for each application.",
-    searchTerms: ["intro", "profile", "overview"],
-  },
-  {
-    id: "picture",
-    label: "Picture",
-    description: "Resume photo and picture presentation.",
-    searchTerms: ["photo", "avatar", "image"],
-  },
-  {
-    id: "basics-custom-fields",
-    label: "Custom Fields",
-    description: "Extra links or short details near your contact info.",
-    searchTerms: ["links", "custom", "details"],
-  },
-];
-
-const DESIGN_RESUME_ICON_GROUPS: DesignResumeNavGroup[] = [
-  {
-    id: "preview",
-    label: "Preview",
-    items: [
-      {
-        id: "live-preview",
-        label: "Live preview",
-        description: "See a preview of your resume as you edit it.",
-        icon: Eye,
-        sectionId: null,
-      },
-    ],
-  },
-  {
-    id: "profile",
-    label: "Profile",
-    items: [
-      {
-        id: "basics",
-        label: "Contact",
-        description: "Name, headline, and contact details.",
-        icon: UserRound,
-      },
-      {
-        id: "summary",
-        label: "Summary",
-        description:
-          "Short intro shown near the top of your resume. Rewritten during Job Tailoring for each application.",
-        icon: FileText,
-      },
-      {
-        id: "picture",
-        label: "Picture",
-        description: "Resume photo and picture presentation.",
-        icon: ImageIcon,
-      },
-      {
-        id: "basics-custom-fields",
-        label: "Custom Fields",
-        description: "Extra links or short details near your contact info.",
-        icon: ListPlus,
-      },
-    ],
-  },
-  {
-    id: "sections",
-    label: "Resume Sections",
-    items: ITEM_DEFINITIONS.map((definition) => ({
-      id: definition.key,
-      label: definition.title,
-      description: definition.description,
-      icon: SECTION_ICON_BY_ID[definition.key] ?? BookOpen,
-    })),
-  },
-];
-
-const DESIGN_RESUME_NAV_GROUPS: SectionWorkspaceGroup<
-  DesignResumeGroupId,
-  DesignResumeSectionId
->[] = [
-  {
-    id: "profile",
-    label: "Profile",
-    items: DESIGN_RESUME_PROFILE_SECTIONS,
-  },
-  {
-    id: "sections",
-    label: "Resume Sections",
-    items: ITEM_DEFINITIONS.map((definition) => ({
-      id: definition.key,
-      label: definition.title,
-      description: definition.description,
-      searchTerms: [
-        definition.singularTitle,
-        definition.primaryField,
-        definition.secondaryField ?? "",
-      ].filter(Boolean),
-    })),
-  },
-];
-
-function getSectionWorkspaceCopy(
-  activeSection: DesignResumeSectionId | null,
-  draft: DesignResumeDocument | null,
-) {
-  const activeSectionMeta = activeSection
-    ? allDesignResumeSections.find((item) => item.id === activeSection)
-    : null;
-  if (!activeSectionMeta) return null;
-
-  if (activeSection !== "basics-custom-fields" || !draft) {
-    return activeSectionMeta;
-  }
-
-  const resumeJson = draft.resumeJson as Record<string, unknown>;
-  const basics = asRecord(resumeJson.basics) ?? {};
-  const customFieldsTitle =
-    toText(basics.customFieldsTitle).trim() || activeSectionMeta.label;
-
-  return {
-    ...activeSectionMeta,
-    label: customFieldsTitle,
-  };
-}
-
-const allDesignResumeSections = DESIGN_RESUME_NAV_GROUPS.flatMap(
-  (group) => group.items,
-);
-const DESIGN_RESUME_ICON_ITEM_BY_SECTION_ID = new Map(
-  DESIGN_RESUME_ICON_GROUPS.flatMap((group) =>
-    group.items.map((item) => [
-      item.sectionId === undefined ? item.id : item.sectionId,
-      item,
-    ]),
-  ),
-);
-
-function getDesignResumeSectionIcon(sectionId: DesignResumeSectionId) {
-  return DESIGN_RESUME_ICON_ITEM_BY_SECTION_ID.get(sectionId)?.icon ?? BookOpen;
-}
-
-function bucketResumeSectionItemCounts(document: DesignResumeJson): {
-  sectionCountBucket: string;
-  itemCountBucket: string;
-} {
-  const sections = asRecord(document.sections) ?? {};
-  const sectionCount = Object.keys(sections).length;
-  const itemCount = Object.values(sections).reduce<number>((count, section) => {
-    const sectionRecord = asRecord(section) ?? {};
-    return count + asArray(sectionRecord.items).length;
-  }, 0);
-  return {
-    sectionCountBucket: bucketCount(sectionCount),
-    itemCountBucket: bucketCount(itemCount),
-  };
-}
-
-function getImportFileType(file: File): "pdf" | "docx" | "unknown" {
-  const mimeType = file.type.toLowerCase();
-  if (mimeType === "application/pdf") return "pdf";
-  if (
-    mimeType ===
-    "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
-  ) {
-    return "docx";
-  }
-  const fileName = file.name.toLowerCase();
-  if (fileName.endsWith(".pdf")) return "pdf";
-  if (fileName.endsWith(".docx")) return "docx";
-  return "unknown";
-}
-
-function getDeviceLayout(): "mobile" | "desktop" {
-  if (typeof window === "undefined" || typeof window.matchMedia !== "function")
-    return "desktop";
-  return window.matchMedia("(max-width: 639px)").matches ? "mobile" : "desktop";
-}
-
-const useDockItemSize = (
-  mouseY: MotionValue<number>,
-  baseItemSize: number,
-  magnification: number,
-  distance: number,
-  ref: React.RefObject<HTMLButtonElement | null>,
-  spring: { mass: number; stiffness: number; damping: number },
-) => {
-  const mouseDistance = useTransform(mouseY, (value) => {
-    if (typeof value !== "number" || Number.isNaN(value)) return 0;
-    const rect = ref.current?.getBoundingClientRect() ?? {
-      y: 0,
-      height: baseItemSize,
-    };
-    return value - rect.y - baseItemSize / 2;
-  });
-
-  const targetSize = useTransform(
-    mouseDistance,
-    [-distance, 0, distance],
-    [baseItemSize, magnification, baseItemSize],
-  );
-
-  return useSpring(targetSize, spring);
-};
-
-type DesignResumeDockItem = {
-  id: string;
-  icon: React.ReactNode;
-  label: string;
-  active: boolean;
-  onClick: () => void;
-  badgeCount?: number;
-  isReorderable?: boolean;
-  isDragging?: boolean;
-  isDragTarget?: boolean;
-  isCandidate?: boolean;
-  onDragStart?: (event: React.DragEvent<HTMLButtonElement>) => void;
-  onDragEnd?: () => void;
-  onDragOver?: (event: React.DragEvent<HTMLButtonElement>) => void;
-  onDrop?: (event: React.DragEvent<HTMLButtonElement>) => void;
-};
-
-type DesignResumeDockButtonProps = DesignResumeDockItem & {
-  mouseY: MotionValue<number>;
-  baseItemSize: number;
-  magnification: number;
-  distance: number;
-  spring: { mass: number; stiffness: number; damping: number };
-};
-
-function DesignResumeDockButton({
-  icon,
-  label,
-  active,
-  onClick,
-  mouseY,
-  baseItemSize,
-  magnification,
-  distance,
-  spring,
-  badgeCount,
-  isReorderable,
-  isDragging,
-  isDragTarget,
-  isCandidate,
-  onDragStart,
-  onDragEnd,
-  onDragOver,
-  onDrop,
-}: DesignResumeDockButtonProps) {
-  const ref = useRef<HTMLButtonElement>(null);
-  const size = useDockItemSize(
-    mouseY,
-    baseItemSize,
-    magnification,
-    distance,
-    ref,
-    spring,
-  );
-
-  return (
-    <motion.button
-      layout
-      transition={{
-        type: "spring",
-        stiffness: 400,
-        damping: 30,
-      }}
-      onClick={onClick}
-      draggable={isReorderable}
-      // biome-ignore lint/suspicious/noExplicitAny: Framer Motion onDragStart type conflicts with native HTML5 drag events
-      onDragStart={onDragStart as any}
-      // biome-ignore lint/suspicious/noExplicitAny: Framer Motion onDragEnd type conflicts with native HTML5 drag events
-      onDragEnd={onDragEnd as any}
-      onDragOver={onDragOver}
-      onDrop={onDrop}
-      className="group relative flex shrink-0 items-center justify-center cursor-grab"
-    >
-      <Tooltip>
-        <TooltipTrigger asChild>
-          <button
-            type="button"
-            ref={ref}
-            className="flex items-center justify-center ml-4 -translate-x-2"
-          >
-            {isReorderable && (
-              <GripVertical
-                className="absolute -left-4 h-3.5 w-3.5 text-muted-foreground/60 opacity-0 transition-opacity duration-150 group-hover:opacity-100 pointer-events-none"
-                aria-hidden="true"
-              />
-            )}
-
-            <motion.button
-              type="button"
-              style={{ width: size, height: size }}
-              animate={{
-                scale: isDragging
-                  ? 0.9
-                  : isDragTarget
-                    ? 1.1
-                    : isCandidate
-                      ? 0.96
-                      : 1,
-              }}
-              transition={{
-                type: "spring",
-                stiffness: 300,
-                damping: 25,
-              }}
-              className={cn(
-                "relative inline-flex cursor-grab shrink-0 items-center justify-center rounded-full border bg-background text-muted-foreground shadow-md outline-none transition-colors duration-150 hover:text-foreground focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 focus-visible:ring-offset-background",
-                isDragTarget
-                  ? "border-primary shadow-primary/30 border-solid"
-                  : isCandidate
-                    ? "border-dashed border-primary/45 bg-primary/5 text-primary/80 shadow-sm shadow-primary/5"
-                    : active
-                      ? "border-primary/50 bg-primary/12 text-primary shadow-primary/20"
-                      : "border-border/70 hover:border-border hover:bg-accent/70",
-              )}
-              aria-current={active ? "page" : undefined}
-              aria-label={label}
-            >
-              <span className="[&_svg]:h-5 [&_svg]:w-5 transition-all duration-150">
-                {icon}
-              </span>
-              {badgeCount !== undefined && badgeCount > 0 ? (
-                <span className="absolute -right-1.5 -top-1.5 flex h-5 min-w-5 items-center justify-center rounded-full bg-primary px-1 text-[10px] font-bold leading-none text-primary-foreground">
-                  {badgeCount > 99 ? "99+" : badgeCount}
-                </span>
-              ) : null}
-            </motion.button>
-          </button>
-        </TooltipTrigger>
-
-        <TooltipContent
-          side="right"
-          sideOffset={12}
-          className="border border-border/70 bg-popover px-2 py-1 text-xs font-medium text-popover-foreground shadow-lg"
-        >
-          {label}
-        </TooltipContent>
-      </Tooltip>
-    </motion.button>
-  );
-}
-
-type DesignResumeIconRailProps = {
-  activeSectionId: DesignResumeSectionId | null;
-  onSectionSelect: (sectionId: DesignResumeSectionId | null) => void;
-  className?: string;
-  draft: DesignResumeDocument | null;
-  onUpdateResumeJson: (
-    updater: (resumeJson: DesignResumeJson) => DesignResumeJson,
-  ) => void;
-};
-
-function DesignResumeDock({
-  activeSectionId,
-  onSectionSelect,
-  className,
-  draft,
-  onUpdateResumeJson,
-}: DesignResumeIconRailProps) {
-  const railRef = useRef<HTMLElement>(null);
-  const contentRef = useRef<HTMLDivElement>(null);
-  const mouseY = useMotionValue(Number.POSITIVE_INFINITY);
-  const spring = { mass: 0.1, stiffness: 150, damping: 12 };
-  const panelWidth = 76;
-  const magnification = 64;
-  const baseItemSize = 46;
-  const distance = 200;
-  const railPadding = 12;
-  const [scrollOffset, setScrollOffset] = useState(0);
-  const [railHeight, setRailHeight] = useState(0);
-  const [contentHeight, setContentHeight] = useState(0);
-
-  const [draggingKey, setDraggingKey] = useState<string | null>(null);
-  const [dragOverKey, setDragOverKey] = useState<string | null>(null);
-
-  const resetDragState = useCallback(() => {
-    setDraggingKey(null);
-    setDragOverKey(null);
-  }, []);
-
-  const handleDragStart = useCallback(
-    (event: React.DragEvent<HTMLButtonElement>, key: string) => {
-      setDraggingKey(key);
-      setDragOverKey(key);
-      event.dataTransfer.effectAllowed = "move";
-      event.dataTransfer.setData("text/plain", key);
-    },
-    [],
-  );
-
-  const handleDragEnd = useCallback(() => {
-    resetDragState();
-  }, [resetDragState]);
-
-  const handleDragOver = useCallback(
-    (event: React.DragEvent<HTMLButtonElement>, key: string) => {
-      if (draggingKey == null) return;
-      if (
-        !REORDERABLE_SECTION_KEYS.includes(draggingKey) ||
-        !REORDERABLE_SECTION_KEYS.includes(key)
-      ) {
-        return;
-      }
-      event.preventDefault();
-      event.dataTransfer.dropEffect = "move";
-      setDragOverKey(key);
-    },
-    [draggingKey],
-  );
-
-  const handleDrop = useCallback(
-    (event: React.DragEvent<HTMLButtonElement>, targetKey: string) => {
-      event.preventDefault();
-      const sourceKey = event.dataTransfer.getData("text/plain") || draggingKey;
-      if (!sourceKey || sourceKey === targetKey) {
-        resetDragState();
-        return;
-      }
-
-      if (
-        !REORDERABLE_SECTION_KEYS.includes(sourceKey) ||
-        !REORDERABLE_SECTION_KEYS.includes(targetKey)
-      ) {
-        resetDragState();
-        return;
-      }
-
-      onUpdateResumeJson((current) => {
-        const next = structuredClone(current);
-        const currentOrder = getSectionOrder(next as Record<string, unknown>);
-        const fromIndex = currentOrder.indexOf(sourceKey);
-        const toIndex = currentOrder.indexOf(targetKey);
-        if (fromIndex !== -1 && toIndex !== -1) {
-          const nextOrder = [...currentOrder];
-          const [removed] = nextOrder.splice(fromIndex, 1);
-          nextOrder.splice(toIndex, 0, removed);
-
-          if (!next.metadata) {
-            next.metadata = {} as DesignResumeJson["metadata"];
-          }
-          if (!next.metadata.layout) {
-            next.metadata.layout = { sidebarWidth: 0, pages: [] };
-          }
-          if (!next.metadata.layout.pages) {
-            next.metadata.layout.pages = [];
-          }
-          if (next.metadata.layout.pages.length === 0) {
-            next.metadata.layout.pages.push({
-              fullWidth: false,
-              main: nextOrder,
-              sidebar: [],
-            });
-          } else {
-            next.metadata.layout.pages[0].main = nextOrder;
-          }
-        }
-        return next;
-      });
-
-      trackProductEvent("resume_studio_section_edited", {
-        section: sourceKey,
-        action: "reorder",
-        item_count_bucket: "0",
-        device_layout:
-          typeof window !== "undefined" &&
-          window.matchMedia?.("(max-width: 639px)").matches
-            ? "mobile"
-            : "desktop",
-      });
-
-      resetDragState();
-    },
-    [draggingKey, onUpdateResumeJson, resetDragState],
-  );
-
-  const items = useMemo(() => {
-    return DESIGN_RESUME_ICON_GROUPS.flatMap((group) => {
-      let groupItems = group.items;
-      if (group.id === "sections" && draft) {
-        const order = getSectionOrder(
-          draft.resumeJson as Record<string, unknown>,
-        );
-        const reorderable = groupItems.filter((item) =>
-          REORDERABLE_SECTION_KEYS.includes(item.id),
-        );
-        reorderable.sort((a, b) => order.indexOf(a.id) - order.indexOf(b.id));
-
-        let reorderableIndex = 0;
-        groupItems = groupItems.map((item) => {
-          if (REORDERABLE_SECTION_KEYS.includes(item.id)) {
-            return reorderable[reorderableIndex++];
-          }
-          return item;
-        });
-      }
-
-      return groupItems.map((item) => {
-        const Icon = item.icon;
-        const sectionId =
-          item.sectionId === undefined ? item.id : item.sectionId;
-        const isReorderable = REORDERABLE_SECTION_KEYS.includes(item.id);
-        const isCandidate =
-          draggingKey !== null && isReorderable && draggingKey !== item.id;
-        return {
-          id: item.id,
-          icon: <Icon aria-hidden="true" />,
-          label: item.label,
-          active: sectionId === activeSectionId,
-          onClick: () => onSectionSelect(sectionId),
-          isReorderable,
-          isDragging: draggingKey === item.id,
-          isDragTarget: dragOverKey === item.id && draggingKey !== item.id,
-          isCandidate,
-          onDragStart: isReorderable
-            ? (e: React.DragEvent<HTMLButtonElement>) =>
-                handleDragStart(e, item.id)
-            : undefined,
-          onDragEnd: isReorderable ? handleDragEnd : undefined,
-          onDragOver: isReorderable
-            ? (e: React.DragEvent<HTMLButtonElement>) =>
-                handleDragOver(e, item.id)
-            : undefined,
-          onDrop: isReorderable
-            ? (e: React.DragEvent<HTMLButtonElement>) => handleDrop(e, item.id)
-            : undefined,
-        };
-      });
-    });
-  }, [
-    draft,
-    activeSectionId,
-    onSectionSelect,
-    draggingKey,
-    dragOverKey,
-    handleDragStart,
-    handleDragEnd,
-    handleDragOver,
-    handleDrop,
-  ]);
-
-  const maxScrollOffset = Math.max(
-    0,
-    contentHeight + railPadding * 2 - railHeight,
-  );
-
-  useEffect(() => {
-    const rail = railRef.current;
-    const content = contentRef.current;
-    if (!rail || !content) return;
-
-    const updateRailHeight = () => setRailHeight(rail.clientHeight);
-    const updateContentHeight = () => setContentHeight(content.scrollHeight);
-    updateRailHeight();
-    updateContentHeight();
-
-    const resizeObserver = new ResizeObserver(() => {
-      updateRailHeight();
-      updateContentHeight();
-    });
-    resizeObserver.observe(rail);
-    resizeObserver.observe(content);
-    return () => resizeObserver.disconnect();
-  }, []);
-
-  useEffect(() => {
-    setScrollOffset((current) => Math.min(current, maxScrollOffset));
-  }, [maxScrollOffset]);
-
-  const handleWheel = (event: React.WheelEvent<HTMLElement>) => {
-    if (maxScrollOffset <= 0) return;
-    event.preventDefault();
-    setScrollOffset((current) =>
-      Math.min(Math.max(current + event.deltaY, 0), maxScrollOffset),
-    );
-  };
-
-  return (
-    <div
-      style={{ width: panelWidth }}
-      className={cn(
-        "pointer-events-none z-30 flex items-start justify-end justify-self-end",
-        className,
-      )}
-    >
-      <TooltipProvider delayDuration={0} skipDelayDuration={80}>
-        <motion.nav
-          ref={railRef}
-          onMouseMove={({ clientY }) => {
-            mouseY.set(clientY);
-          }}
-          onMouseLeave={() => {
-            mouseY.set(Number.POSITIVE_INFINITY);
-          }}
-          onBlur={() => mouseY.set(Number.POSITIVE_INFINITY)}
-          onWheel={handleWheel}
-          className="pointer-events-auto relative h-full min-h-0 w-[64px] overflow-x-visible overflow-y-clip overscroll-contain rounded-2xl border border-border/80 bg-card/95 shadow-2xl shadow-background/50 backdrop-blur supports-[backdrop-filter]:bg-card/85"
-          role="toolbar"
-          aria-label="Resume Studio sections"
-        >
-          <motion.div
-            ref={contentRef}
-            className="absolute left-0 right-0 top-3 flex flex-col items-center gap-2"
-            style={{ y: -scrollOffset }}
-          >
-            {items.map((item) => (
-              <DesignResumeDockButton
-                key={item.id}
-                {...item}
-                mouseY={mouseY}
-                baseItemSize={baseItemSize}
-                magnification={magnification}
-                distance={distance}
-                spring={spring}
-              />
-            ))}
-          </motion.div>
-        </motion.nav>
-      </TooltipProvider>
-    </div>
-  );
-}
 
 export const DesignResumePage: React.FC = () => {
-  const queryClient = useQueryClient();
   const navigate = useNavigate();
-  const { section: sectionParam } = useParams<{ section?: string }>();
-  const { document, status, isLoading, error } = useDesignResume();
-  const { settings, isLoading: settingsLoading } = useSettings();
-  const { readiness: tracerReadiness } = useTracerReadiness();
-  const [draft, setDraft] = useState<DesignResumeDocument | null>(null);
-  const [saveState, setSaveState] = useState<
-    "idle" | "saving" | "saved" | "error"
-  >("idle");
-  const [dialogState, setDialogState] = useState<{
-    definition: ItemDefinition;
-    index: number | null;
-    seed: Record<string, unknown> | null;
-  } | null>(null);
-  const [pictureUploading, setPictureUploading] = useState(false);
-  const [resumeImporting, setResumeImporting] = useState(false);
-  const [showReimportConfirm, setShowReimportConfirm] = useState(false);
-  const [mobileSectionPickerOpen, setMobileSectionPickerOpen] = useState(false);
-  const [mobileWorkspaceView, setMobileWorkspaceView] =
-    useState<DesignResumeMobileView>(() => (sectionParam ? "edit" : "preview"));
-  const [pdfDownloading, setPdfDownloading] = useState(false);
-  const [rendererUpdating, setRendererUpdating] = useState(false);
-  const [dirty, setDirty] = useState(false);
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const importFileInputRef = useRef<HTMLInputElement>(null);
-  const editVersionRef = useRef(0);
-  const draftRef = useRef<DesignResumeDocument | null>(null);
-  const readyPdfRefreshToastShownRef = useRef(false);
-  draftRef.current = draft;
-
-  const notifyReadyPdfRefresh = useCallback(() => {
-    if (readyPdfRefreshToastShownRef.current) return;
-    readyPdfRefreshToastShownRef.current = true;
-    toast.info("Ready PDFs will refresh automatically.");
-  }, []);
-
-  const pdfRenderer = settings?.pdfRenderer?.value ?? "rxresume";
-  const typstTheme = settings?.typstTheme?.value ?? "classic";
-  const canDownloadPdf = status?.exists && !pdfDownloading;
-  const pictureEnabled = Boolean(tracerReadiness?.isPubliclyAvailable);
-  const pictureDisabledReason =
-    tracerReadiness?.reason ??
-    "Pictures require JobOps to be reachable at a public URL.";
-  const activeSection = sectionParam ?? null;
-  const activeSectionIsValid =
-    activeSection == null ||
-    allDesignResumeSections.some((item) => item.id === activeSection);
-
-  useEffect(() => {
-    setMobileWorkspaceView(sectionParam ? "edit" : "preview");
-    setMobileSectionPickerOpen(false);
-  }, [sectionParam]);
-
-  useEffect(() => {
-    if (!document) return;
-    setDraft(document);
-    setDirty(false);
-  }, [document]);
-
-  useEffect(() => {
-    if (
-      !draft ||
-      !document ||
-      !dirty ||
-      saveState === "saving" ||
-      saveState === "error"
-    ) {
-      return;
-    }
-
-    const timer = window.setTimeout(async () => {
-      const editVersionAtStart = editVersionRef.current;
-      const baseRevision = draft.revision;
-      const documentSnapshot = structuredClone(draft.resumeJson);
-
-      try {
-        setSaveState("saving");
-        const updated = await api.updateDesignResume({
-          baseRevision,
-          document: documentSnapshot,
-        });
-        if (editVersionRef.current === editVersionAtStart) {
-          queryClient.setQueryData(queryKeys.designResume.current(), updated);
-          queryClient.setQueryData(queryKeys.designResume.status(), {
-            exists: true,
-            documentId: updated.id,
-            updatedAt: updated.updatedAt,
-          });
-          setDraft(updated);
-          setDirty(false);
-          setSaveState("saved");
-          notifyReadyPdfRefresh();
-          return;
-        }
-
-        // Keep any newer local edits, but advance the base revision for the
-        // next autosave cycle so stale responses never clobber in-flight work.
-        setDraft((current) =>
-          current
-            ? {
-                ...updated,
-                resumeJson: current.resumeJson,
-              }
-            : updated,
-        );
-        setSaveState("idle");
-      } catch (saveError) {
-        setSaveState("error");
-        showErrorToast(saveError, "Failed to save Resume Studio.");
-      }
-    }, 700);
-
-    return () => window.clearTimeout(timer);
-  }, [dirty, draft, document, notifyReadyPdfRefresh, queryClient, saveState]);
-
-  const setDesignResume = (next: DesignResumeDocument) => {
-    queryClient.setQueryData(queryKeys.designResume.current(), next);
-    queryClient.setQueryData(queryKeys.designResume.status(), {
-      exists: true,
-      documentId: next.id,
-      updatedAt: next.updatedAt,
-    });
-    setDraft(next);
-    setDirty(false);
-  };
-
-  const ensureLatestPersistedDraft =
-    async (): Promise<DesignResumeDocument | null> => {
-      if (!draft) return null;
-      if (!dirty) return draft;
-      if (saveState === "saving") {
-        throw new Error(
-          "Resume Studio is still saving. Try again in a moment.",
-        );
-      }
-
-      const editVersionAtStart = editVersionRef.current;
-      const baseRevision = draft.revision;
-      const documentSnapshot = structuredClone(draft.resumeJson);
-
-      setSaveState("saving");
-      const updated = await api.updateDesignResume({
-        baseRevision,
-        document: documentSnapshot,
-      });
-
-      if (editVersionRef.current === editVersionAtStart) {
-        setDesignResume(updated);
-        setSaveState("saved");
-        return updated;
-      }
-
-      const mergedResumeJson =
-        draftRef.current?.resumeJson ?? updated.resumeJson;
-      const mergedDraft = {
-        ...updated,
-        resumeJson: structuredClone(mergedResumeJson) as DesignResumeJson,
-      };
-      setDraft((current) =>
-        current
-          ? {
-              ...updated,
-              resumeJson: current.resumeJson,
-            }
-          : updated,
-      );
-      setDirty(true);
-      setSaveState("idle");
-      return mergedDraft;
-    };
-
-  const updateResumeJson = (
-    updater: (resumeJson: DesignResumeJson) => DesignResumeJson,
-  ) => {
-    editVersionRef.current += 1;
-    setDraft((current) => {
-      if (!current) return current;
-      return {
-        ...current,
-        resumeJson: updater(current.resumeJson),
-      };
-    });
-    setDirty(true);
-    if (saveState === "saved" || saveState === "error") setSaveState("idle");
-  };
-
-  const activeDialogItem = useMemo(() => {
-    if (!dialogState) return null;
-    return (
-      dialogState.seed ??
-      (dialogState.index == null
-        ? dialogState.definition.createItem()
-        : getDesignResumeDialogItem(
-            draft,
-            dialogState.definition,
-            dialogState.index,
-          ))
-    );
-  }, [dialogState, draft]);
-
-  const handleImport = async () => {
-    const wasReimport = Boolean(status?.exists);
-    try {
-      setResumeImporting(true);
-      const imported = await api.importDesignResumeFromRxResume();
-      setDesignResume(imported);
-      setSaveState("saved");
-      const counts = bucketResumeSectionItemCounts(imported.resumeJson);
-      trackProductEvent("resume_studio_import_completed", {
-        source: "rxresume",
-        result: "success",
-        was_reimport: wasReimport,
-        section_count_bucket: counts.sectionCountBucket,
-        item_count_bucket: counts.itemCountBucket,
-      });
-      if (!wasReimport) {
-        trackProductEvent("resume_studio_activation_completed", {
-          source: "rxresume",
-        });
-      }
-      toast.success("Imported your resume.");
-      notifyReadyPdfRefresh();
-    } catch (importError) {
-      trackProductEvent("resume_studio_import_completed", {
-        source: "rxresume",
-        result: "error",
-        was_reimport: wasReimport,
-        section_count_bucket: "0",
-        item_count_bucket: "0",
-      });
-      showErrorToast(importError, "Failed to import your resume.");
-    } finally {
-      setResumeImporting(false);
-    }
-  };
-
-  const handleImportWithConfirm = () => {
-    if (status?.exists) {
-      setShowReimportConfirm(true);
-    } else {
-      void handleImport();
-    }
-  };
-
-  const handleImportFile = async (file: File) => {
-    const wasReimport = Boolean(status?.exists);
-    const fileType = getImportFileType(file);
-    try {
-      setResumeImporting(true);
-      const dataUrl = await fileToDataUrl(file);
-      const match = /^data:([^;]+);base64,(.+)$/s.exec(dataUrl.trim());
-
-      if (!match) {
-        throw new Error("Resume file could not be encoded for upload.");
-      }
-
-      const imported = await api.importDesignResumeFromFile({
-        fileName: file.name,
-        mediaType: file.type || match[1],
-        dataBase64: match[2],
-      });
-      setDesignResume(imported);
-      setSaveState("saved");
-      const counts = bucketResumeSectionItemCounts(imported.resumeJson);
-      trackProductEvent("resume_studio_import_completed", {
-        source: "file",
-        file_type: fileType,
-        result: "success",
-        was_reimport: wasReimport,
-        section_count_bucket: counts.sectionCountBucket,
-        item_count_bucket: counts.itemCountBucket,
-      });
-      if (!wasReimport) {
-        trackProductEvent("resume_studio_activation_completed", {
-          source: "file",
-        });
-      }
-      toast.success("Imported your resume file.");
-      notifyReadyPdfRefresh();
-    } catch (importError) {
-      setSaveState("error");
-      trackProductEvent("resume_studio_import_completed", {
-        source: "file",
-        file_type: fileType,
-        result: "error",
-        was_reimport: wasReimport,
-        section_count_bucket: "0",
-        item_count_bucket: "0",
-      });
-      showErrorToast(importError, "Failed to import your resume file.");
-    } finally {
-      setResumeImporting(false);
-      if (importFileInputRef.current) {
-        importFileInputRef.current.value = "";
-      }
-    }
-  };
-
-  const handleExport = async () => {
-    try {
-      const exported = await api.exportDesignResume();
-      makeDownload(exported.fileName, exported.document);
-      trackProductEvent("resume_studio_export_completed", {
-        result: "success",
-      });
-      toast.success("Exported your resume JSON.");
-    } catch (exportError) {
-      trackProductEvent("resume_studio_export_completed", { result: "error" });
-      showErrorToast(exportError, "Failed to export Resume Studio.");
-    }
-  };
-
-  const handleDownloadPdf = async () => {
-    const downloadedAfterEdit = dirty || saveState === "saving";
-    try {
-      setPdfDownloading(true);
-      const generated = await api.generateDesignResumePdf();
-      await downloadDesignResumePdf(generated.fileName, generated.pdfUrl);
-      trackProductEvent("resume_studio_pdf_downloaded", {
-        renderer: pdfRenderer,
-        theme: typstTheme,
-        after_edit: downloadedAfterEdit,
-        result: "success",
-      });
-      toast.success("Your PDF is ready.");
-    } catch (downloadError) {
-      trackProductEvent("resume_studio_pdf_downloaded", {
-        renderer: pdfRenderer,
-        theme: typstTheme,
-        after_edit: downloadedAfterEdit,
-        result: "error",
-      });
-      showErrorToast(downloadError, "Failed to generate a PDF.");
-    } finally {
-      setPdfDownloading(false);
-    }
-  };
-
-  const handleUploadPicture = async (file: File) => {
-    if (!pictureEnabled) {
-      toast.error(pictureDisabledReason);
-      if (fileInputRef.current) {
-        fileInputRef.current.value = "";
-      }
-      return;
-    }
-
-    try {
-      setPictureUploading(true);
-      const latestDraft = await ensureLatestPersistedDraft();
-      if (!latestDraft) return;
-
-      const editVersionAtStart = editVersionRef.current;
-      const updated = await api.uploadDesignResumePictureFile({
-        file,
-        baseRevision: latestDraft.revision,
-      });
-      if (editVersionRef.current === editVersionAtStart) {
-        setDesignResume(updated);
-      } else {
-        setDraft((current) =>
-          current
-            ? {
-                ...updated,
-                resumeJson: current.resumeJson,
-              }
-            : updated,
-        );
-        setDirty(true);
-        setSaveState("idle");
-      }
-      toast.success("Picture uploaded.");
-      notifyReadyPdfRefresh();
-    } catch (uploadError) {
-      showErrorToast(uploadError, "Failed to upload picture.");
-    } finally {
-      setPictureUploading(false);
-      if (fileInputRef.current) {
-        fileInputRef.current.value = "";
-      }
-    }
-  };
-
-  const handleDeletePicture = async () => {
-    try {
-      const latestDraft = await ensureLatestPersistedDraft();
-      if (!latestDraft) return;
-
-      const editVersionAtStart = editVersionRef.current;
-      const updated = await api.deleteDesignResumePicture({
-        baseRevision: latestDraft.revision,
-        document: latestDraft.resumeJson,
-      });
-      if (editVersionRef.current === editVersionAtStart) {
-        setDesignResume(updated);
-      } else {
-        setDraft((current) =>
-          current
-            ? {
-                ...updated,
-                resumeJson: current.resumeJson,
-              }
-            : updated,
-        );
-        setDirty(true);
-        setSaveState("idle");
-      }
-      toast.success("Picture removed.");
-      notifyReadyPdfRefresh();
-    } catch (deleteError) {
-      showErrorToast(deleteError, "Failed to delete picture.");
-    }
-  };
-
-  const handlePdfRendererChange = async (nextRenderer: PdfRenderer) => {
-    if (settingsLoading || nextRenderer === pdfRenderer) return;
-
-    try {
-      setRendererUpdating(true);
-      const updatedSettings = await api.updateSettings({
-        pdfRenderer: nextRenderer,
-      });
-      queryClient.setQueryData(queryKeys.settings.current(), updatedSettings);
-      toast.success(`${PDF_RENDERER_LABELS[nextRenderer]} is now active.`);
-      notifyReadyPdfRefresh();
-    } catch (updateError) {
-      showErrorToast(updateError, "Failed to update the resume template.");
-    } finally {
-      setRendererUpdating(false);
-    }
-  };
-
-  const handleTypstThemeChange = async (nextTheme: TypstTheme) => {
-    if (settingsLoading || nextTheme === typstTheme) return;
-
-    try {
-      setRendererUpdating(true);
-      const updatedSettings = await api.updateSettings({
-        typstTheme: nextTheme,
-      });
-      queryClient.setQueryData(queryKeys.settings.current(), updatedSettings);
-      toast.success(`${TYPST_THEME_LABELS[nextTheme]} Typst theme is active.`);
-      notifyReadyPdfRefresh();
-    } catch (updateError) {
-      showErrorToast(updateError, "Failed to update the Typst theme.");
-    } finally {
-      setRendererUpdating(false);
-    }
-  };
-
-  const activeSectionMeta = getSectionWorkspaceCopy(activeSection, draft);
-  const activeGroup = activeSection
-    ? DESIGN_RESUME_NAV_GROUPS.find((group) =>
-        group.items.some((item) => item.id === activeSection),
-      )
-    : null;
-
-  const handleMobileSectionSelect = (sectionId: DesignResumeSectionId) => {
-    setMobileWorkspaceView("edit");
-    setMobileSectionPickerOpen(false);
-    navigate(`/design-resume/${sectionId}`);
-  };
-
-  const getDesignResumeSectionBadge = useCallback(
-    (sectionId: DesignResumeSectionId): SectionWorkspaceBadge | null => {
-      if (!draft) return null;
-      const resumeJson = draft.resumeJson as Record<string, unknown>;
-      if (sectionId === "basics") {
-        const basics = asRecord(resumeJson.basics) ?? {};
-        return toText(basics.name) || toText(basics.headline)
-          ? { label: "Ready", variant: "outline" }
-          : { label: "Empty", variant: "secondary" };
-      }
-      if (sectionId === "summary") {
-        const summary = asRecord(resumeJson.summary) ?? {};
-        return toText(summary.content)
-          ? { label: "Ready", variant: "outline" }
-          : { label: "Empty", variant: "secondary" };
-      }
-      if (sectionId === "picture") {
-        const picture = asRecord(resumeJson.picture) ?? {};
-        return toText(picture.url)
-          ? { label: "Uploaded", variant: "outline" }
-          : { label: "Optional", variant: "secondary" };
-      }
-      if (sectionId === "basics-custom-fields") {
-        const basics = asRecord(resumeJson.basics) ?? {};
-        const count = asArray(basics.customFields).length;
-        return {
-          label: count === 0 ? "Empty" : `${count}`,
-          variant: "secondary",
-        };
-      }
-
-      const sections = asRecord(resumeJson.sections) ?? {};
-      const section = asRecord(sections[sectionId]) ?? {};
-      const count = asArray(section.items).length;
-      return {
-        label: count === 0 ? "Empty" : `${count}`,
-        variant: count === 0 ? "secondary" : "outline",
-      };
-    },
-    [draft],
-  );
+  const {
+    draft,
+    dirty,
+    saveState,
+    dialogState,
+    pictureUploading,
+    resumeImporting,
+    showReimportConfirm,
+    mobileSectionPickerOpen,
+    mobileWorkspaceView,
+    pdfDownloading,
+    rendererUpdating,
+    settingsLoading,
+    isLoading,
+    error,
+    status,
+    activeSection,
+    activeSectionIsValid,
+    activeSectionMeta,
+    activeGroup,
+    activeDialogItem,
+    pdfRenderer,
+    typstTheme,
+    canDownloadPdf,
+    pictureEnabled,
+    pictureDisabledReason,
+    fileInputRef,
+    importFileInputRef,
+    setDialogState,
+    setShowReimportConfirm,
+    setMobileSectionPickerOpen,
+    setMobileWorkspaceView,
+    updateResumeJson,
+    handleImport,
+    handleImportWithConfirm,
+    handleImportFile,
+    handleExport,
+    handleDownloadPdf,
+    handleUploadPicture,
+    handleDeletePicture,
+    handlePdfRendererChange,
+    handleTypstThemeChange,
+    handleMobileSectionSelect,
+    getDesignResumeSectionBadge,
+  } = useDesignResumeStudio();
 
   if (!activeSectionIsValid) {
     return <Navigate to="/design-resume" replace />;
